requirements.txt
asgiref==3.8.1
attrs==23.2.0
betterpath==0.2.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
cssselect==1.2.0
d==0.2.2
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.0.4
django-allauth==0.61.1
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
l==0.11.0
lxml==5.2.1
Markdown==3.6
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pyquery==2.0.0
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.35.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
vcversioner==2.16.0.0
zope.interface==6.3
uwsgi
